Seperated two crons in two different service files
