+++ src/src/cron/cron.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
-import { CronService } from './cron.service';
+import { FaceIndexingService } from './faceIndexing.service';
+import { MarkAttendanceService } from './markAttendance.service';
 import { AwsRekognitionModule } from '../services/aws-rekognition/aws-rekognition.module';
 import { HasuraModule } from '../services/hasura/hasura.module';
 @Module({
 	imports: [AwsRekognitionModule, HasuraModule],
-	providers: [CronService],
+	providers: [FaceIndexingService, MarkAttendanceService],
 })
 export class CronModule {}
